Split block_graph.py into smaller files
