+++ src/tqec/sketchup/block_graph/graph.py
@@ -5,285 +5,20 @@ from __future__ import annotations
 import pathlib
 import typing as ty
 from copy import deepcopy
-from dataclasses import astuple, dataclass
-from enum import Enum
 from io import BytesIO
 
 import networkx as nx
 
 from tqec.exceptions import TQECException
 from tqec.position import Direction3D, Position3D
+from tqec.sketchup.block_graph.cube import Cube
+from tqec.sketchup.block_graph.enums import CubeType, PipeType
+from tqec.sketchup.block_graph.observable import AbstractObservable
+from tqec.sketchup.block_graph.pipe import Pipe
 from tqec.sketchup.zx_graph import NodeType, ZXGraph, ZXNode
 
 if ty.TYPE_CHECKING:
     from tqec.sketchup.collada import ColladaDisplayHelper
-
-
-class Color(Enum):
-    X = "x"
-    Z = "z"
-    NULL = "o"
-
-    @property
-    def is_null(self) -> bool:
-        """Check if the color is null."""
-        return self == Color.NULL
-
-
-@dataclass(frozen=True)
-class Color3D:
-    """Get face colors along the x, y, and z axes."""
-
-    x: Color
-    y: Color
-    z: Color
-
-    @staticmethod
-    def null() -> Color3D:
-        """Get the null color."""
-        return Color3D(Color.NULL, Color.NULL, Color.NULL)
-
-    @property
-    def is_null(self) -> bool:
-        """Check if the color is null."""
-        return self == Color3D.null()
-
-    def match(self, other: Color3D) -> bool:
-        """Check whether the color matches the other color."""
-        return all(
-            c1.is_null or c2.is_null or c1 == c2
-            for c1, c2 in zip(astuple(self), astuple(other))
-        )
-
-    def pop_color_at_direction(self, direction: Direction3D) -> Color3D:
-        """Replace the color at the given direction with None."""
-        return self.push_color_at_direction(direction, Color.NULL)
-
-    def push_color_at_direction(self, direction: Direction3D, color: Color) -> Color3D:
-        """Set the color at the given direction."""
-        colors = list(astuple(self))
-        colors[direction.axis_index] = color
-        return Color3D(*colors)
-
-    @staticmethod
-    def from_string(s: str, flip_xz: bool = False) -> Color3D:
-        s = s.lower()
-        if s == "virtual":
-            return Color3D.null()
-        if len(s) != 3 or any(c not in "xzo" for c in s):
-            raise TQECException(
-                "s must be a 3-character string containing only 'x', 'z', and 'o'."
-            )
-        colors: list[Color] = []
-        for c in s:
-            if c == "o":
-                colors.append(Color.NULL)
-            elif flip_xz:
-                colors.append(Color.Z if c == "x" else Color.X)
-            else:
-                colors.append(Color(c))
-        return Color3D(*colors)
-
-
-class CubeType(Enum):
-    """Valid cube types in the library."""
-
-    ZXX = "zxx"
-    XZX = "xzx"
-    XXZ = "xxz"
-    XZZ = "xzz"
-    ZXZ = "zxz"
-    ZZX = "zzx"
-    # Virtual cube for open port
-    VIRTUAL = "virtual"
-
-    @property
-    def is_spatial_junction(self) -> bool:
-        return self in [CubeType.ZZX, CubeType.XXZ]
-
-    def to_zx_node_type(self) -> NodeType:
-        """Convert the cube type to a ZX node type."""
-        if self == CubeType.VIRTUAL:
-            return NodeType.V
-        elif self.value.count("z") == 2:
-            return NodeType.Z
-        return NodeType.X
-
-    def get_color(self) -> Color3D:
-        """Get the color of the block."""
-        return Color3D.from_string(self.value)
-
-    @staticmethod
-    def from_color(color: Color3D) -> CubeType:
-        """Get the cube type from the color."""
-        if color.is_null:
-            return CubeType.VIRTUAL
-        if any(c.is_null for c in astuple(color)):
-            raise TQECException("All the color must be defined for a non-virtual cube.")
-        return CubeType("".join(c.value for c in astuple(color)).lower())
-
-    def normal_direction_to_corner_plane(self) -> Direction3D:
-        """If the cube is at a corner, return the normal direction to the
-        corner plane.
-
-        Due to the color match rule at the corner turn, the corner plane
-        can be inferred from the type of the cube.
-        """
-        if self == CubeType.VIRTUAL:
-            raise TQECException("Cannot infer the corner plane for a virtual cube.")
-        if self.value.count("z") == 2:
-            return Direction3D.from_axis_index(self.value.index("x"))
-        return Direction3D.from_axis_index(self.value.index("z"))
-
-    def infer_pipe_type_at_direction(
-        self,
-        direction: Direction3D,
-        src_side_if_h_pipe: bool = True,
-        has_hadamard: bool = False,
-    ) -> PipeType:
-        """Infer the pipe type connecting this cube at some direction with the
-        color match rule."""
-        if self == CubeType.VIRTUAL:
-            raise TQECException("Cannot infer the pipe type for a virtual cube.")
-        color = self.get_color().pop_color_at_direction(direction)
-        return PipeType.from_color_at_side(color, src_side_if_h_pipe, has_hadamard)
-
-
-class PipeType(Enum):
-    """Valid pipe types in the library."""
-
-    # pipes without H
-    OZX = "ozx"
-    OXZ = "oxz"
-    XOZ = "xoz"
-    ZOX = "zox"
-    XZO = "xzo"
-    ZXO = "zxo"
-    # pipes with H
-    OZXH = "ozxh"
-    OXZH = "oxzh"
-    ZOXH = "zoxh"
-    XOZH = "xozh"
-    ZXOH = "zxoh"
-    XZOH = "xzoh"
-
-    @property
-    def has_hadamard(self) -> bool:
-        """Check if the block type has a hadamard transition."""
-        return "h" in self.value
-
-    @property
-    def direction(self) -> Direction3D:
-        """Return the direction of the pipe."""
-        return Direction3D.all()[self.value.index("o")]
-
-    def get_color_at_side(self, src_side_if_h_pipe: bool = True) -> Color3D:
-        """Get the color of the pipe at the given side."""
-        if not self.has_hadamard or src_side_if_h_pipe:
-            return Color3D.from_string(self.value[:3])
-        return Color3D.from_string(self.value[:3], flip_xz=True)
-
-    @staticmethod
-    def from_color_at_side(
-        color: Color3D, src_side_if_h_pipe: bool = True, has_hadamard: bool = False
-    ) -> PipeType:
-        """Get the pipe type from the color at one side."""
-        if not sum(c.is_null for c in astuple(color)) == 1:
-            raise TQECException(
-                "Exactly one color must be undefined for a pipe along the pipe direction."
-            )
-        pipe_color = []
-        for c in astuple(color):
-            if c.is_null:
-                pipe_color.append("o")
-            elif has_hadamard and not src_side_if_h_pipe:
-                pipe_color.append("x" if c == Color.Z else "z")
-            else:
-                pipe_color.append(c.value)
-        if has_hadamard:
-            pipe_color.append("h")
-        return PipeType("".join(pipe_color).lower())
-
-    def infer_cube_type_at_side(
-        self, src_side_if_h_pipe: bool = True, is_z_cube: bool = True
-    ) -> CubeType:
-        """Infer the cube type at the side of the pipe."""
-        color = self.get_color_at_side(src_side_if_h_pipe).push_color_at_direction(
-            self.direction, Color.Z if is_z_cube else Color.X
-        )
-        return CubeType.from_color(color)
-
-    def temporal_basis(self) -> ty.Literal["X", "Z"]:
-        if self.direction == Direction3D.Z:
-            raise TQECException("Cannot get the temporal basis on a temporal pipe.")
-        # Because this is not a temporal pipe, we know for sure
-        # that there is no "o" in the Z direction. Also, "h" can
-        # not appear in the first 3 entries, so self.value[2] is
-        # either "x" or "z".
-        return ty.cast(ty.Literal["X", "Z"], self.value[2].upper())
-
-
-BlockType = ty.Union[CubeType, PipeType]
-"""Valid block types in the library."""
-
-
-@dataclass(frozen=True)
-class Cube:
-    """A block representing the computational unit in a 3D spacetime
-    diagram."""
-
-    position: Position3D
-    cube_type: CubeType
-
-    @property
-    def is_virtual(self) -> bool:
-        """Check if the cube is virtual, i.e. an open port."""
-        return self.cube_type == CubeType.VIRTUAL
-
-
-@dataclass(frozen=True)
-class Pipe:
-    """A block connecting two cubes in a 3D spacetime diagram.
-
-    The pipe represents the idle or merge/split lattice surgery
-    operation on logical qubits depending on its direction.
-    """
-
-    u: Cube
-    v: Cube
-    pipe_type: PipeType
-
-    def __post_init__(self) -> None:
-        if not self.u.position.is_neighbour(self.v.position):
-            raise TQECException("A pipe must connect two nearby cubes.")
-        # Ensure position of u is less than v
-        u, v = self.u, self.v
-        if self.u.position > self.v.position:
-            object.__setattr__(self, "u", v)
-            object.__setattr__(self, "v", u)
-
-    @property
-    def direction(self) -> Direction3D:
-        """Get the direction of the pipe."""
-        return self.pipe_type.direction
-
-
-@dataclass(frozen=True)
-class AbstractObservable:
-    """An abstract description of an observable in a 3D spacetime diagram.
-
-    In a **closed** 3D spacetime diagram, the abstract observable can be derived from
-    the corresponding correlation surface:
-    1. When the correlation surface attaches to the top/bottom faces of a block in
-    the diagram, the measurements of the line of qubits on the top face are included
-    in the observable.
-    2. When the correlation surface lies within XY plane and intersects a pipe, the stabilizer
-    measurements at the start of the pipe and part of the stabilizer measurements within
-    the cubes connected by the pipe are included in the observable.
-    """
-
-    top_lines: frozenset[Cube | Pipe]
-    bottom_regions: frozenset[Pipe]
 
 
 _CUBE_DATA_KEY = "tqec_block_cube_data"
